Added an update of the time in the Site table when traversing each new page. I have completed the startIndexing method, which must respond with different response code.

+++ searchengine-master/src/main/java/searchengine/services/IndexingServiceImpl.java
@@ -6,11 +6,8 @@ import org.springframework.stereotype.Service;
 import searchengine.config.Site;
 import searchengine.config.SitesList;
 import searchengine.dto.statistics.MyThreadParser;
-import searchengine.model.Status;
 import searchengine.repository.PageRepository;
 import searchengine.repository.SiteRepository;
-
-import java.time.LocalDateTime;
 import java.util.*;
 
 @Service
@@ -28,21 +25,28 @@ public class IndexingServiceImpl implements IndexingService{
     @Override
     public Map<String,String> startedIndexing() {
 
-        Map<String, String> map = new HashMap<>();
         List<Site> listSites = sites.getSites();
-        String[] result = { "true", "false" };
-        String error = "Индексация уже запущена";
+        List<Thread> threadList = new ArrayList<>();
+        Map<String, String> map = new HashMap<>();
 
-        siteRepository.deleteAll();
-        pageRepository.deleteAll();
-
-        for(Site site : listSites){
-            MyThreadParser myThreadParser = new MyThreadParser(pageRepository, siteRepository, site);
-            Thread myThread = new Thread(myThreadParser);
-            myThread.start();
+        if(MyThreadParser.getPoolSize() > 0){
+            Map<String, String> map1 = new HashMap<>();
+            map1.put("result", "false");
+            map1.put("error", "Индексация уже запущена");
+            return map1;
         }
 
-        map.put("result", result[0]);
+        if(!MyThreadParser.fjpIsActive()){
+            for(Site site : listSites){
+                siteRepository.deleteAll();
+                pageRepository.deleteAll();
+                MyThreadParser myThreadParser = new MyThreadParser(pageRepository, siteRepository, site);
+                Thread myThread = new Thread(myThreadParser);
+                threadList.add(myThread);
+                myThread.start();
+            }
+        }
+        map.put("result", "true");
         return map;
     }
